# oba-ui/src/app/api/swagger/oba.json
{
	"swagger": "2.0",
	"info": {
		"description": "Implementation of backend for online banking UI. We have 3 preloaded users in Ledgers: <b>marion.mueller</b>, <b>anton.brueckner</b>, <b>max.musterman</b> all with the PIN <b>12345</b>. You can use the User Login API from Ledgers <b>/users/authorise2</b> endpoint to gain an access token. Then use the access token with the prefix 'Bearer ' to Authorize on this ui.",
		"version": "0.5.0",
		"title": "Online banking",
		"termsOfService": "Terms of Service: to be edited...",
		"contact": {
			"name": "OFIN",
			"url": "https://www.ofin.co",
			"email": "fpo@ofin.co"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "https://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "dev-dynamicsandbox-onlinebanking.cloud.ofin.co",
	"basePath": "/",
...
	"paths": {
...
		"/oauth/authorise": {
			"post": {
				"tags": [
					"Online Banking Oauth Authorization"
				],
				"summary": "Oauth authorise",
				"operationId": "oauthCodeUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "login",
						"in": "header",
						"description": "login",
						"required": true,
						"type": "string"
					},
					{
						"name": "pin",
						"in": "header",
						"description": "pin",
						"required": true,
						"type": "string"
					},
					{
						"name": "redirect_uri",
						"in": "query",
						"description": "redirect_uri",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/OauthCodeResponseTO"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/oauth/authorization-server": {
			"get": {
				"tags": [
					"Online Banking Oauth Authorization"
				],
				"summary": "Server info",
				"operationId": "oauthServerInfoUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cancellationId",
						"in": "query",
						"description": "cancellationId",
						"required": false,
						"type": "string"
					},
					{
						"name": "consentId",
						"in": "query",
						"description": "consentId",
						"required": false,
						"type": "string"
					},
					{
						"name": "paymentId",
						"in": "query",
						"description": "paymentId",
						"required": false,
						"type": "string"
					},
					{
						"name": "redirectId",
						"in": "query",
						"description": "redirectId",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/OauthServerInfoTO"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/oauth/token": {
			"post": {
				"tags": [
					"Online Banking Oauth Authorization"
				],
				"summary": "Oauth token",
				"operationId": "oauthTokenUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "code",
						"in": "query",
						"description": "code",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/BearerTokenTO"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
	},
	"securityDefinitions": {
		"apiKey": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"definitions": {
		"AccessTokenTO": {
			"type": "object",
			"properties": {
				"act": {
					"type": "object",
					"description": "The bearer this token.",
					"additionalProperties": {
						"type": "string"
					}
				},
				"authorisation_id": {
					"type": "string",
					"description": "The last authorisation id leading to this token"
				},
				"consent": {
					"description": "The specification of psd2 account access permission associated with this token",
					"$ref": "#/definitions/AisConsentRequest"
				},
				"exp": {
					"type": "string",
					"format": "date-time",
					"description": "expiration time"
				},
				"iat": {
					"type": "string",
					"format": "date-time",
					"description": "Issue time"
				},
				"jti": {
					"type": "string",
					"description": "The token identifier"
				},
				"login": {
					"type": "string",
					"description": "The login name of the initiator of this token"
				},
				"role": {
					"type": "string",
					"description": "Role to be inforced when this token is presented.",
					"enum": [
						"CUSTOMER",
						"STAFF",
						"TECHNICAL",
						"SYSTEM"
					]
				},
				"sca_id": {
					"type": "string",
					"description": "The id of the sca object: login, payment, account access"
				},
				"sub": {
					"type": "string",
					"description": "The database id of the initiator of this token"
				},
				"token_usage": {
					"type": "string",
					"description": "The usage of this token.",
					"enum": [
						"LOGIN",
						"DIRECT_ACCESS",
						"DELEGATED_ACCESS"
					]
				}
			},
			"title": "AccessTokenTO",
			"description": "The access token object."
		},
...
		"AisAccountAccessInfo": {
			"type": "object",
			"properties": {
				"accounts": {
					"type": "array",
					"description": "Access to accounts",
					"items": {
						"type": "string"
					}
				},
				"allPsd2": {
					"type": "string",
					"example": "ALL_ACCOUNTS",
					"description": "Consent on all accounts, balances and transactions of psu",
					"enum": [
						"ALL_ACCOUNTS",
						"ALL_ACCOUNTS_WITH_BALANCES"
					]
				},
				"availableAccounts": {
					"type": "string",
					"example": "ALL_ACCOUNTS",
					"description": "Consent on all available accounts of psu",
					"enum": [
						"ALL_ACCOUNTS",
						"ALL_ACCOUNTS_WITH_BALANCES"
					]
				},
				"balances": {
					"type": "array",
					"description": "Access to balances",
					"items": {
						"type": "string"
					}
				},
				"transactions": {
					"type": "array",
					"description": "Access to transactions",
					"items": {
						"type": "string"
					}
				}
			},
			"title": "AisAccountAccessInfo",
			"description": "Ais account access information"
		},
...
		"AisConsentRequest": {
			"type": "object",
			"required": [
				"access",
				"frequencyPerDay",
				"id",
				"recurringIndicator",
				"tppId",
				"userId",
				"validUntil"
			],
			"properties": {
				"access": {
					"description": "Set of accesses given by psu for this account",
					"$ref": "#/definitions/AisAccountAccessInfo"
				},
				"frequencyPerDay": {
					"type": "integer",
					"format": "int32",
					"example": 4,
					"description": "Maximum frequency for an access per day. For a once-off access, this attribute is set to 1"
				},
				"id": {
					"type": "string",
					"description": "The consent id"
				},
				"recurringIndicator": {
					"type": "boolean",
					"example": false,
					"description": "'true', if the consent is for recurring access to the account data , 'false', if the consent is for one access to the account data"
				},
				"tppId": {
					"type": "string",
					"example": "testTPP",
					"description": "ID of the corresponding TPP."
				},
				"userId": {
					"type": "string",
					"description": "Corresponding PSU"
				},
				"validUntil": {
					"type": "string",
					"format": "date",
					"example": "2020-10-10",
					"description": "Consent`s expiration date. The content is the local ASPSP date in ISODate Format"
				}
			},
			"title": "AisConsentRequest",
			"description": "Ais consent request"
		},
...
		"BearerTokenTO": {
			"type": "object",
			"properties": {
				"accessTokenObject": {
					"$ref": "#/definitions/AccessTokenTO"
				},
				"access_token": {
					"type": "string"
				},
				"expires_in": {
					"type": "integer",
					"format": "int32"
				},
				"refresh_token": {
					"type": "string"
				},
				"token_type": {
					"type": "string"
				}
			},
			"title": "BearerTokenTO"
		},
...
		"OauthCodeResponseTO": {
			"type": "object",
			"properties": {
				"redirectUri": {
					"type": "string"
				}
			},
			"title": "OauthCodeResponseTO"
		},
		"OauthServerInfoTO": {
			"type": "object",
			"properties": {
				"authorization_endpoint": {
					"type": "string"
				},
				"grant_types_supported": {
					"type": "array",
					"items": {
						"type": "string",
						"enum": [
							"AUTHORISATION_CODE",
							"REFRESH_TOKEN"
						]
					}
				},
				"response_types_supported": {
					"type": "array",
					"items": {
						"type": "string",
						"enum": [
							"CODE"
						]
					}
				},
				"token_endpoint": {
					"type": "string"
				}
			},
			"title": "OauthServerInfoTO"
		},
...
	}
}